fix(folder-tree): hide graph-collapse icon for unloaded folders
A newly-created folder is unloaded by default and has no graph presence,
yet the sidebar's graph-collapse icon was rendered for every folder with a
graphFolderId. Since the icon's collapsed/expanded class is a pure binary
derived from the collapsed-folder set, an unloaded folder (never in that
set) painted the "expanded" class while appearing nowhere in the graph —
the "expanded yet invisible" contradiction.

The three-state folder visibility {expanded, collapsed, hidden} is flattened
to a binary collapseSet at the server→client projection boundary
(application/session/project.ts:316), so the client cannot tell 'hidden'
(unloaded) from 'expanded'. The sidebar already distinguishes load state via
a separate load indicator, so the correct fix is to gate the graph-collapse
affordance on it: an unloaded folder has no collapse/expand state to show.

Changes:
- FolderTreeNode.tsx: render the graph-collapse icon only when
  node.loadState === 'loaded' (implements BF-114's "skips … unloaded" intent).
- FolderTreeNode.test.tsx: unit regression — a not-loaded folder renders no
  graph-collapse icon (7/7 pass).
- Tier-3 black-box electron e2e: an unloaded child folder shows no
  graph-collapse icon, and gains one once loaded
  (electron-filetree-unloaded-folder-no-graph-collapse.spec.ts).
- Extracted the shared load-child electron fixture into
  electron-filetree-load-child.fixtures.ts (consumed by both folder-tree
  load specs) to avoid duplicating the launch boilerplate.
